modify: replace Levenshtein with SorencenDice for role suggestions
