+++ build.gradle.kts
@@ -44,6 +44,7 @@ dependencies {
     implementation("dev.misfitlabs.kotlinguice4:kotlin-guice:1.4.1")
 
     implementation("net.imagej:imagej:2.2.0")
+    implementation("info.debatty:java-string-similarity:2.0.0")
 
     testImplementation(kotlin("test-junit5"))
     testImplementation("org.junit.jupiter:junit-jupiter-api:5.6.0")
